requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
--index-url http://edx.devstack.devpi:3141/root/pypi/+simple/
--trusted-host edx.devstack.devpi

astroid==1.6.6            # via pylint
backports.functools-lru-cache==1.5  # via astroid, isort, pylint
bumpversion==0.5.3
certifi==2019.3.9         # via requests
chardet==3.0.4            # via requests
configparser==3.7.4       # via pylint
django==1.11.15
edx-opaque-keys==0.4.4
enum34==1.1.6             # via astroid
futures==3.2.0            # via isort
idna==2.8                 # via requests
isort==4.3.17             # via pylint
lazy-object-proxy==1.3.1  # via astroid
mccabe==0.6.1             # via pylint
pbr==5.1.3                # via stevedore
pycodestyle==2.4.0
pylint==1.9.3
pymongo==3.7.2            # via edx-opaque-keys
pytz==2018.9              # via django
requests==2.21.0          # via rocketchat-api
rocketchat_api==0.6.21
singledispatch==3.4.0.3   # via astroid, pylint
six==1.12.0               # via astroid, edx-opaque-keys, pylint, singledispatch, stevedore
stevedore==1.30.1         # via edx-opaque-keys
urllib3==1.24.1           # via requests
wrapt==1.11.1             # via astroid
